Move log message test support code to separate file
These classes might be re-used by many tests.
